ARM64: use 64-bit operands for mrs/msr inline assembly
mrs/msr only take 64-bit registers, but several snippets pass 32-bit
variables. GCC prints the x register anyway, clang refuses:

  arch/arm/include/asm/system.h:112:43: error: value size does not match register size specified by the constraint and modifier [-Werror,-Wasm-operand-widths]

Make the asm operands 64-bit; the function signatures are unchanged.

+++ arch/arm/cpu/atf.c
@@ -4,7 +4,7 @@
 #include <asm/atf_common.h>
 #include <asm/system.h>
 
-static inline void raw_write_daif(unsigned int daif)
+static inline void raw_write_daif(unsigned long daif)
 {
 	__asm__ __volatile__("msr DAIF, %0\n\t" : : "r" (daif) : "memory");
 }
